# Clase 13/Actividad en mesas/docs/swagger.yaml
definitions:
  domain.Product:
    properties:
      code_value:
        type: string
      expiration:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    required:
    - code_value
    - expiration
    - name
    - price
    - quantity
    type: object
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact:
    name: API Support
    url: https://developers.ctd.com.ar/support
  description: This API Handle Products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.ctd.com.ar/es_ar/terminos-y-condiciones
  title: Certified Tech Developer
  version: "1.0"
paths:
  /products:
    get:
      description: Gets all the products from the repository
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Gets all the products
      tags:
      - products
    post:
      description: Create a new product in repository
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product from the repository
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Deletes a product
      tags:
      - products
    get:
      description: Gets a product by id from the repository
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Gets a product by id
      tags:
      - products
    patch:
      description: Updates selected fields from a product from the repository
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Product
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Updates selected fields
      tags:
      - products
    put:
      description: Updates a product from the repository
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Product
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Updates a product
      tags:
      - products
  /products/consumer_price:
    get:
      description: Returns the prices of one or more products from the repository
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Returns the prices of one or more products
      tags:
      - products
  /products/search:
    get:
      description: Search for a product by price from the repository
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Search for a product by price
      tags:
      - products
swagger: "2.0"